Jonah wip register artefact (#4)
* Add react icons package fortawesome. Fix category creation not setting state to include new category. WIP fixing the categorySelect not adding the newly created category to the list of selected vals

* Fixed category select not adding newly created categories to the list of selected vals

Still an issue if a user adds a category and then deletes and then readds. This appears to be because react-select maintains it's own internal state, which overrides the prop 'categoryOpts' passed to it (possibly since the select option has the same value). There is an 'onCreate' prop that would be useful here, a function of form \(nameOfNewOption). This would require updating the options and values list of the react-select comp., which did not seem to happen by simply modifying the 'option' and 'values' props - :(

* Login / signup pages

* Remove appsettings.json for azure config

* Add azuredb connection to startup and migrations