ImageCarousel behaviour changes.
Make slide transition quicker. The original transition speed was sooooo slow,
especially for testing.

Remove displaying of placeholder when image list is empty. The placeholder
clutters up the scroller when many artefacts do not have images, and makes
the site look tacky and/or broken (even if it is technically functioning
correctly).