Fix frontend update issues with editing artefact images.
We need to update the state from props every render, because the images are
changed outside the ArtefactInfo component, unlike all the other parts, because
they are controlled by a modal window, and bootstrap modal windows don't work
inside positioned components, which ArtefactInfo is.

Is it too late to transition to an Arts major?